refactor: move Vulkan texture import logic from GPURenderingContext to a dedicated module
