+++ examples/servo/src/webview/rendering_context/servo_rendering_adapter.rs
@@ -59,11 +59,9 @@ struct ServoGPURenderingContext {
 impl ServoRenderingAdapter for ServoGPURenderingContext {
     fn current_framebuffer_as_image(&self) -> Image {
         #[cfg(any(target_os = "linux", target_os = "android"))]
-        let texture = self.rendering_context
-            .get_wgpu_texture_from_vulkan(&self.device, &self.queue)
-            .expect(
-                "Failed to get WGPU texture from Vulkan texture - ensure rendering context is valid",
-            );
+        let texture = self.rendering_context.get_wgpu_texture_from_vulkan(&self.device).expect(
+            "Failed to get WGPU texture from Vulkan texture - ensure rendering context is valid",
+        );
 
         #[cfg(target_vendor = "apple")]
         let texture =
